+++ addon/query/odata-adapter.js
@@ -6,6 +6,7 @@ import { SimplePredicate, ComplexPredicate, StringPredicate, DetailPredicate } f
 import FilterOperator from './filter-operator';
 import Information from '../utils/information';
 import getSerializedDateValue from '../utils/get-serialized-date-value';
+import { classify } from '../utils/string-functions';
 
 /**
  * Class of query adapter that translates query object into OData URL.
@@ -330,7 +331,7 @@ export default class ODataAdapter extends BaseAdapter {
           Ember.warn(`Source type is not specified for the enum '${meta.type}' (${modelName}.${predicate.attributePath}).`,
           false,
           { id: 'ember-flexberry-data-debug.odata-adapter.source-type-is-not-specified-for-enum' });
-          type = Ember.String.classify(meta.type);
+          type = classify(meta.type);
         }
 
         value = `${type}'${predicate.value}'`;